SubSample uses float instead of int for time and charge. Make using this consistent throughout the tools Create a sorting algorithm in TriggerInfo. Add ability to add other TriggerInfo instances to the current one In DataOut, save digits in subevents into new WCSimRootTrigger instances Add logging to nhits Add adjust_for_noise option in nhits
Apparently issues don't exist on TriggerApplication github... so I put them here...
Add a int version of time/charge in SubSample for GPU triggers to use
Implement an option to save only 1 digit per pmt per trigger (this is how SKI worked)

+++ UserTools/WCSimASCIReader/WCSimASCIReader.cpp
@@ -22,7 +22,7 @@ bool WCSimASCIReader::Execute(){
   std::ifstream data (inputfile.c_str());
   if (data.is_open()){
     std::vector<int>PMTid;
-    std::vector<int>time;
+    std::vector<float>time;
     while ( getline (data,line) )
       {
 	int tmpPMTid=0;
